# Avoid char array allocation in Starlark `format`