@fmeum 2024-11-25
Thanks, I fixed and reran the benchmark. It now shows that the new implementation is only better with the Latin-1 string encoding, but worse in UTF-8 standalone Starlark. Not sure whether that makes it a good idea.