@tetromino 2024-11-25
Interesting. In that case, we may want different behavior depending on whether the string is Latin1...