@fmeum 2025-10-16
@tetromino I could gate this behind the semantics flag introduced in https://github.com/bazelbuild/bazel/pull/24417.