@tetromino 2025-10-16
> @tetromino I could gate this behind the semantics flag introduced in #24417.

@fmeum - good idea!